bugfixes and training of 128 64 for 120 epochs. Fixed the val loss to use the prediction scores as loss (1 - pred)
