+++ pcdet/models/detectors/detector3d_template.py
@@ -362,10 +362,8 @@ class Detector3DTemplate(nn.Module):
         checkpoint = torch.load(filename, map_location=loc_type)
         epoch = checkpoint.get('epoch', -1)
         it = checkpoint.get('it', 0.0)
-
         self.load_state_dict(checkpoint['model_state'])
-
-        if optimizer is not None:
+        if optimizer is not None:  
             if 'optimizer_state' in checkpoint and checkpoint['optimizer_state'] is not None:
                 logger.info('==> Loading optimizer parameters from checkpoint %s to %s'
                             % (filename, 'CPU' if to_cpu else 'GPU'))
