+++ tools/cfgs/kitti_models/centerpoint.yaml
@@ -49,7 +49,7 @@ MODEL:
 
     POST_PROCESSING:
         RECALL_THRESH_LIST: [0.3, 0.5, 0.7]
-        SCORE_THRESH: 0.1  # 0.1 original value, made higher for real inference and reduce number of FPs. It shows scores higher than this
+        SCORE_THRESH: 0.3 # 0.1 original value, made higher for real inference and reduce number of FPs. It shows scores higher than this
         OUTPUT_RAW_SCORE: False
 
         EVAL_METRIC: kitti
